refactor(image listing page): modularize functions and components
